fix: Allow null origin for map tile requests
