+++ api.planx.uk/modules/ordnanceSurvey/routes.ts
@@ -1,8 +1,9 @@
 import { Router } from "express";
 import { useOrdnanceSurveyProxy } from "./controller.js";
+import { osCors } from "../../cors.js";
 
 const router = Router();
 
-router.use("/proxy/ordnance-survey", useOrdnanceSurveyProxy);
+router.use("/proxy/ordnance-survey", osCors, useOrdnanceSurveyProxy);
 
 export default router;
